requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
google-generativeai>=0.3.1
python-multipart>=0.0.6

pydantic~=2.10.4
google~=3.0.0
protobuf~=5.29.2